[REF] Moved helper methods to utils and proper message format.
- Helper methods that are intended to help in some ways where moved to the proper utils lib.

- Json message is better formed to accomplish the previous standard that was in the PoC and the receiver.

- Added some missing coding headers
